Better integration with upstream image
Use build-standalone script not to include the standalone build of the
library on the image. It's only useful for autotools-like builds and it
makes it considerably bigger.
Furhtermore, bundle the different builds together so we only include the
additions to /opt we require, without the intermediate steps (binary
builds, source code, etc).